This merge is nowhere near finished: everything's broken!

+++ src/Backends/Rendering/Window/OpenGL.h
@@ -1,5 +1,10 @@
+// Released under the MIT licence.
+// See LICENCE.txt for details.
+
 #pragma once
 
-bool WindowBackend_OpenGL_CreateWindow(const char *window_title, int *screen_width, int *screen_height, bool fullscreen, bool vsync);
+#include <stddef.h>
+
+bool WindowBackend_OpenGL_CreateWindow(const char *window_title, size_t *screen_width, size_t *screen_height, bool fullscreen, bool vsync);
 void WindowBackend_OpenGL_DestroyWindow(void);
 void WindowBackend_OpenGL_Display(void);
